Fix C++ move-constructed function arguments
Due to the prior change, moved arguments were being passed directly
like a value type. However, these should actually be passed by
reference because the caller is responsible for calling the destructor
once the callee returns, and any modifications made by the callee need
to be reflected in the caller's context.