Don't collapse 8-byte memcpys into 64-bit operands
The instcombine optimization pass assumes a 64-bit register is available
and tries to collapse aligned memcpys that are <= 8 bytes into a single
load/store, but this doesn't work on our poor 32-bit CPU and it ends up
producing code that our back-end must compensate for, causing issues.

In the most egregious cases, it was falling back to 64-bit "polyfill"
functions such as __lshrdi3 to support these, which is way slower than
just leaving them alone.

Note that another pass appears to inline small memcpys, so they still
end up getting optimized away.

This PR puts the v810 fork of clang out of sync with the main repo. One
wonders if this should be fixed upstream to query the machine metadata.